backend/requirements.txt
# you will need to install these requirements with `pip install -r requirements.txt`
SQLAlchemy==1.2.11
aiohttp-jinja2==1.0.0
aiohttp-session[secure]==2.5.1
aiohttp==3.4.4
aiopg==0.15.0
pytest-aiohttp==0.3.0
pytest-cov==2.5.1
pytest==3.8.1
aiohttp_cors
gitpython
aiohttp-devtools
isort
mccabe
beautifulsoup4
lxml
